+++ benches/benches.rs
@@ -1,11 +1,12 @@
 use cosmian_crypto_core::{
     asymmetric_crypto::{curve25519::X25519KeyPair, DhKeyPair},
-    entropy::CsRng,
     kdf,
+    reexport::rand_core::SeedableRng,
     symmetric_crypto::{aes_256_gcm_pure::Aes256GcmCrypto, key::Key, Dem},
-    KeyTrait,
+    CsRng, KeyTrait,
 };
 use criterion::{criterion_group, criterion_main, Criterion};
+use rand_chacha::rand_core::RngCore;
 use sha3::{
     digest::{ExtendableOutput, Update, XofReader},
     Shake128,
@@ -15,7 +16,7 @@ use sha3::{
 /// Diffie-Hellman key exchange. This gives an indication on how fast
 /// asymmetric schemes can be.
 fn bench_dh(c: &mut Criterion) {
-    let mut rng = CsRng::new();
+    let mut rng = CsRng::from_entropy();
     let dh_keypair = X25519KeyPair::new(&mut rng);
     c.bench_function(
         "Bench the Group-Scalar multiplication on which is based the Diffie-Helman key exchange",
@@ -27,9 +28,10 @@ fn bench_dh(c: &mut Criterion) {
 const MSG_LENGTH: usize = 2048;
 
 fn bench_symmetric_encryption(c: &mut Criterion) {
-    let mut rng = CsRng::new();
+    let mut rng = CsRng::from_entropy();
     let key = Key::new(&mut rng);
-    let msg = rng.generate_random_bytes::<MSG_LENGTH>();
+    let mut msg = [0; MSG_LENGTH];
+    rng.fill_bytes(&mut msg);
     c.bench_function(
         "Bench the DEM encryption of a 2048-bytes message withou additional data",
         |b| b.iter(|| Aes256GcmCrypto::encrypt(&mut rng, &key, &msg, None).unwrap()),
@@ -37,9 +39,10 @@ fn bench_symmetric_encryption(c: &mut Criterion) {
 }
 
 fn bench_symmetric_decryption(c: &mut Criterion) {
-    let mut rng = CsRng::new();
+    let mut rng = CsRng::from_entropy();
     let key = Key::new(&mut rng);
-    let msg = rng.generate_random_bytes::<MSG_LENGTH>();
+    let mut msg = [0; MSG_LENGTH];
+    rng.fill_bytes(&mut msg);
     let enc = Aes256GcmCrypto::encrypt(&mut rng, &key, &msg, None).unwrap();
     c.bench_function(
         "Bench the DEM decryption of a 2048-bytes message withou additional data",
@@ -53,14 +56,15 @@ fn bench_symmetric_decryption(c: &mut Criterion) {
 /// OS needs to gather enough entropy.
 fn bench_rng_generation(c: &mut Criterion) {
     c.bench_function("Bench the generation of a cryptographic RNG", |b| {
-        b.iter(CsRng::new)
+        b.iter(CsRng::from_entropy)
     });
 }
 
 fn bench_kdf(c: &mut Criterion) {
     const LENGTH: usize = 64;
-    let mut rng = CsRng::new();
-    let ikm = rng.generate_random_bytes::<32>();
+    let mut rng = CsRng::from_entropy();
+    let mut ikm = [0; 32];
+    rng.fill_bytes(&mut ikm);
     c.bench_function(
         "Bench the KDF derivation of a 32-bytes IKM into a 64-bytes key",
         |b| b.iter(|| kdf!(LENGTH, &ikm, b"KDF derivation")),